requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
